pyproject.toml
[project]
name = "mcp-server-drugdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb==1.1.3",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
    "pyyaml>=6.0.2",
]
